# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CheckAuroa-API-Schema",
    "description": "CheckAuroa-API-Schema",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
  
  "/location": {
    
    "get": {
      "summary": "geth the location",
      "description":"",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "locationGet",
      "responses": {
        "200": {
  "description" : "OK"
}
        
        
      }
    } 
    
  } ,
  
  
  "/rating": {
    
    "get": {
      "summary": "get a rating for your location",
      "description":"",
      "x-swagger-router-controller": "Ratings",
      "tags": ["Ratings"],
      "operationId": "getRating",
      "parameters": [
        {
  "name" : "lng",
  "in" : "query",
  "description" : "The lat lng position for calculate the rating",
  "required" : true,
  "type" : "number"
},
        {
  "name" : "lat",
  "in" : "query",
  "description" : "The lat lng position for calculate the rating",
  "required" : true,
  "type" : "number"
},
        {
  "name" : "UTCDateTime",
  "in" : "query",
  "description" : "The lat lng position for calculate the rating",
  "required" : false,
  "type" : "string",
  "default" : "now",
  "format" : "dateTime"
}
        
      ],
      "responses": {
        "200": {
  "description" : "An single rating element",
  "schema" : {
    "$ref" : "#/definitions/Rating"
  }
}
        ,
        "default": {
  "description" : "unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  
  
  "/location": {
    
    "post": {
      "summary": "Updates the location for a spesific device",
      "description":"",
      "x-swagger-router-controller": "Location",
      "tags": ["Location"],
      "operationId": "updateLocation",
      "parameters": [
        {
  "name" : "deviceId",
  "in" : "query",
  "description" : "An id for your individual device",
  "required" : true,
  "type" : "string"
},
        {
  "name" : "deviceType",
  "in" : "query",
  "description" : "The device type code it can be on of [ANDOID, WEB, QT]",
  "required" : true,
  "type" : "string",
  "default" : "WEB",
  "enum" : [ "ANDROID", "QT", "WEB" ]
},
        {
  "name" : "lat",
  "in" : "query",
  "description" : "The lat lng position for calculate the rating",
  "required" : true,
  "type" : "number"
},
        {
  "name" : "lng",
  "in" : "query",
  "description" : "The lat lng position for calculate the rating",
  "required" : true,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "OK"
}
        
        
      }
    } 
    
  } ,
  
  
  "/notification": {
    
    "post": {
      "summary": "Create a new notification",
      "description":"",
      "x-swagger-router-controller": "Notify",
      "tags": ["Notify"],
      "operationId": "createNotification",
      "parameters": [
        {
  "in" : "body",
  "name" : "body",
  "description" : "The lat lng position for calculate the rating",
  "required" : true,
  "schema" : {
    "$ref" : "#/definitions/Notification"
  }
}
        
      ],
      "responses": {
        "201": {
  "description" : "Null response"
}
        ,
        "default": {
  "description" : "unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } ,
    
    "delete": {
      "summary": "deletes a notification",
      "description":"This method is to delete a notofication",
      "x-swagger-router-controller": "Notify",
      "tags": ["Notify"],
      "operationId": "deleteNotification",
      "responses": {
        "200": {
  "description" : "OK"
}
        ,
        "404": {
  "description" : "NOT FOUND"
}
        
        
      }
    } 
    
  } 
  
  }, "definitions": {
  "Error": {
  "properties" : {
    "msg" : {
      "type" : "string"
    },
    "code" : {
      "type" : "string",
      "description" : "The error code",
      "default" : "INTERNAL",
      "enum" : [ "INTERNAL", "KP_SERVICE_ERROR", "WEATHER_SERVICE_ERROR", "SUNSETRISE_SERVICE_ERROR" ]
    }
  }
},"Location": {
  "properties" : {
    "lat" : {
      "type" : "number"
    },
    "lng" : {
      "type" : "number"
    }
  }
},"Rating": {
  "properties" : {
    "location" : {
      "$ref" : "#/definitions/Location"
    },
    "kpIndex" : {
      "type" : "integer"
    }
  }
},"Notification": {
  "properties" : {
    "playSound" : {
      "type" : "boolean"
    }
  }
}
  }
}